fix: move key generation header in another folder
