+++ src/pages/ValidatorOnboarding/KeyGeneration/KeyGenerationHeader/KeyGenerationHeader.tsx
@@ -1,6 +1,6 @@
 import { XStack } from 'tamagui'
 import KeyGenerationSyncCard from './KeyGenerationSyncCard'
-import KeyGenerationTitle from './KeyGenerationTitle'
+import KeyGenerationTitle from '../KeyGenerationTitle'
 
 const KeyGenerationHeader = () => {
   return (
